- Progressing in Ubuntu 24.04 update. Building up to ulisse_ctrl. - TODO: Fix Warnings in SafetyBoundaries.cpp and errors in control_test.cpp namespacing

+++ ulisse_driver/src/ulisse_driver/thread_receiver.cpp
@@ -280,7 +280,7 @@ namespace llc {
                 llc_sw485_pub_->publish(llc_sw485_msg_);
                 break;
             default:
-                RCLCPP_WARN(this->get_logger(), "Unhandled message type %d", llcData_.messageType);
+                RCLCPP_WARN(this->get_logger(), "Unhandled message type %d", static_cast<int>(llcData_.messageType));
                 break;
             }
             std::this_thread::sleep_for(1ms);
